[OM] FreezePaths: only resolve paths up to the original path
This changes FreezePaths to only follow paths up to the root of the original
hierarchical path operation, instead of continuing upwards in the instance
graph.  We no longer enforce that the path is unambiguous up to the top level
(or public) module. This is in preparation to support paths when there is
multiple module instantiation, where we will enfore that paths are rooted at
their owning module.