Disable failing sim tests to unblock LLVM bump
We currently don't rely on `llhd-sim` for anything in-tree. The upstream
change to `!llvm.ptr` has triggered a lot of churn in the LLHD side of
CIRCT. A few breakages remain in the interaction between the llhd-sim
runtime and the generated code. These are now gated behind a
`REQUIRES: llhd-sim-fixed` in order to unblock the LLVM bump and allow
for them to be fixed in parallel.