[Debug] Add expression support to HGLDD emission (#6334)
Extend `EmitHGLDD` to properly represent combinational expressions in
the emitted HGLDD debug info file. This is a first prototype to
experiment with recreating higher-level source language values from the
low-level signals that survived all the way through Verilog emission.

Instead of just checking if `dbg.variable` ops directly point at
something in the MLIR that has a public Verilog name, the emitter now
looks through expressions, `dbg.struct`s, and `dbg.array`s, and builds
up the equivalent expression in HGLDD. As a result, the emitted debug
info file now contains enough for a debugger to reconstruct many of the
interesting source language values from the values that are actually
present in the output Verilog.

At a later point we might want to move some of the expression graph
construction into the `DebugInfo` analysis. But at the moment it isn't
totally clear to me whether looking through the IR proper is an
acceptable approach, or whether we'd want to create a "shadow expression
graph" in the analysis that is decoupled from the IR.