WIP: [Arc] Implement Louvain community detection to group ops
Add the `GroupComputation` pass that performs the Louvain method for
community detection on the operations in `arc.clock_tree`s. As a result,
operations are moved into groups such that there is a high amount of SSA
value use within a group, and comparatively few SSA values flowing into
and out of the group.

The implementation uses a topological order to prevent groups from
forming dependency cycles.