[PyCDE] Use bundles for ESI services (#6436)
Fixes all of the PyCDE ESI unit tests. Disable testing anything related to service generation, which needs to be re-thought. Also, ditch packaging and creation of the runtime APIs as this is now the responsibility of the runtime itself (now that we have a separate one). Need to figure out how this is going to interact with PyCDE.