[ExportVerilog] Skip debug dialect ops (#6453)
Skip debug dialect ops in `ExportVerilog` and `PrepareForEmission` since
those operations and values are only used for debug info emission and
cannot be materialized directly in the Verilog output. This allows for
the IR to retain debug operations purely for the purpose of documenting
how Verilog values map back to the source language.

This fixes an issue where enabling debug info in firtool via the `-g`
option causes Verilog emission to fail due to the presence of `dbg.*`
operations.

Ideally we'll use the `DebugAnalysis` in the future to skip entire
expression trees, operations, values, and block arguments if they are
only consumed by debug operations. Doing so isn't as trivial as skipping
ops marked as debug-only by the analysis though, since that also skips
def-before-use legalization and other restructuring that `ExportVerilog`
relies on. Until that is sorted out, skipping ops based on dialect is
a conservative stopgap solution.