[Arc] Mark StateOp as side-effect-free if it has no name
Make `StateOp`s implement the `MemoryEffectsOpInterface` and mark the
operation as side-effect-free if it has no name. Names are the only way
for these ops to become observable for the user. If they have no name,
the user has no way of telling whether a state got deleted or not.