[Arc] Add dead code elimination pass
Add an Arc-specific dead code elimination pass which deletes state and
call ops if they are side-effect-free (not observable by the user) and
not reachable from any output, observable op, or other side-effecting
op.

The pass is very conservative at the moment, deleting only calls and
state ops. If these are mixed with any other ops, even if they are
side-effect-free, they will not be deleted. We may want to be more
aggressive in the future.