[FIRRTL][Parser] Improve rwprobe parsing, support inst results. (#6258)
Parse explicitly for rwprobe case for better handling
and to avoid creating spurious indexing operations.

Instance results are supported by inserting a forceable
bounce wire.  The wire replaces the instance port entirely
(including its unbundled entry) locally, and is connected
to/from the instance result.

Wires are created for entire result (port) only, and will
be re-used by repeated rwprobe's to the result in part or whole.
This means the entire port must be forceable and will go
through the bounce wire.

Annotations will continue to resolve as they do today,
not to the inserted bounce wire.