[Debug] Make paths in HGLDD files relative (#6451)
Make paths emitted into HGLDD companion files relative to the file
itself, or emit absolute paths. For example, the `foo/bar/hello.dd`
companion file to `foo/bar/hello.sv` will now refer to the SV file as
`hello.sv`, and to a `src/input.scala` file as `../../src/input.scala`.
If paths cannot be made relative, for example because the output
directory contains `../` components that cannot be handled, the paths
are made absolute.

This fixes an issue where beforehand a tool consuming the HGLDD file
would have to know the working directory in which `firtool` was run to
create the HGLDD file. Now the files are self-contained, and paths are
anchored at the HGLDD file itself.