[CombToArith] Fix lowering of concat with single operand (#6505)
A `comb.concat` operation with only a single operand would cause an
`arith.extui` to be emitted with the same operand and result type. This
trips the verifier of `extui`. Add a special case to that conversion to
simply remove the `comb.concat` if it only has one operand.

The canonicalizer would usually delete these degenerate concats, which
is why we haven't seen this issue earlier.

Also slightly tweak the lowering to produce a shift and or op for the
least significant operand of the concat, which is already in the right
spot.