[OM] Use type replacer to handle block arguments. (#6532)
We added a type replacer and used it to update the type of list create
ops previously, but a similar issue shows up if there are lists of
paths in block arguments. This factors out some helpers added
previously for list create ops, so the same logic and conversion can
be used in block arguments.