[FIRRTL] Add ref.cast.unsafe
Add an op, ref.cast.unsafe, which can be used for doing un-verified checks
between layer-associated probes.  This is necessary to more cleanly handle
LowerLayers where layerblocks that drive probe-associated layers are
replaced with modules and instances.  However, the information that a
probe is associated with a layer needs to be preserved from LowerLayers
until LowerXMR so that LowerXMR can emit correct ifdef-guarded Verilog
macros.  This provides a clean way to have LwoerXMR handle the emission by
just looking at the type as opposed to having it look this information up
in a custom attribute.

An alternative to this approach would be to use
builtin.unrealized_conversion_cast in place of ref.cast.unsafe.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>