[CombFolds] Flatten operands all at once (#6593)
This fixes O(N^2) compile time regression in OrOp(and other logical operations) canonicalizers. Currently tryFlattenOperands flattens operands by only one level, e.g:

Or(a, Or(b, Or(c, d)))
->
Or(a, b, Or(c, d))
->
Or(a, b, c, d)

Problematically this canonicalizer visits operands until a first operand we can flatten every time so this is problematic when chains is really long. This PR fixes the issue by flattening all operands recursively from a root all at once. Also flatteing is moved to a start of canonicalizations so that other canonicalizations are only applied to root operations. This PR should be essentially NFC but there are several differences in verilog files since canonicalizations are applied in a different order and are not canonical. 

Benchmark for one of the largest internal cores:
```
Before:
  389.7719 (  9.6%)  341.5307 ( 29.6%)    Canonicalizer
  4077.4128 (100.0%)  1153.8959 (100.0%)  Total

After:
   43.4950 (  1.3%)    1.7205 (  0.2%)    Canonicalizer
  3350.8652 (100.0%)  803.7959 (100.0%)  Total
```

Fix one of the problems of https://github.com/llvm/circt/issues/6587