[FIRRTL] Fix shr(<zero-width>, 0) folder crash
Fix a bug in the `shr` folder where a shift by zero of a zero-width value
would cause a crash.  As of FIRRTL 3.3.0 the result width is size 1.  If
this is folded to the input, then it causes a crash because types no
longer match.  Disable this fold, as we do with mux in the same situation.

h/t @fabianschuiki for suggesting the fix.

Fixes #6608.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>