[FIRRTL] Add enabled layers to FModuleOp (#6627)
* [FIRRTL] Add layer colors to FModuleOp

Add the ability for FIRRTL module to enable layers.  This will have the
effect of allowing a module to interact with colored probes of instances
that it instantiates.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>

* [FIRRTL] Add verification of FModuleOp layers

Extend existing verification of 'firrtl.ref.cast', 'firrtl.ref.define',
and 'firrtl.ref.resolve' to handle modules which may enable layers.

This also refactors all the layer-coloring verification into a single
function that is reused by all three operation verifiers.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>

* [FIRRTL] Pipeline updates for module layers

Updates to the FIRRTL pass pipeline to support modules with enabled
layers.  These do not have a real effect on the pipeline itself other than
they should be removed once all ops that rely on their existence to pass
verification are removed.  I.e., this commit only removes them in
LowerXMR.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>

* [FIRRTL] Parse/emit module-enabled layers

Add FIRRTL text parsing and emission for modules with layers enabled.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>

* Update all FModules to support enabled layers

* Add parsing/emitting for enablelayer for Intmodule/Extmodule

- improve formatting of output

---------

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>
Co-authored-by: Robert Young <rwy0717@gmail.com>