[OM] Support integer binary arithmetic in the Evaluator.
This adds support for the Evaluator to evaluate integer binary
arithmetic. This is defined in terms of the IntegerBinaryArithmeticOp
interface, so the Evaluator can generically handle all ops that
implement this interface in terms of the interface methods.

At a high level, getOrCreateValue will create a partially evaluated
value for the result of this op. Once the operands are fully
evaluated, they are passed to the evaluateIntegerOperation interface
method, and if successful, the result is used for the evaluator
value.

Because these are the first kind of operation using AttributeValues
that doesn't necessarily make the Attribute immediately available, the
AttributeValue has been extended to support construction in a
partially evaluated state to be filled in later.

The tests show example of straightforward evaluation where the
operands are immediately available from constants, where the operands
are references to object fields, and where the operands are references
to object fields that are initially not evaluated.

This handles many cases, but there is currently no detection for
dataflow cycles through an integer binary arithmetic operation. The
evaluator would need significant changes in the current setup to
detect this on the fly, so currently assumes cycles are prevented
earlier in the pipeline.