[firtool] Add option to disable black box resource
Add the `--disable-blackbox-resource-file` option to firtool and the
BlackBoxReaderPass.

Fix #4249. Another way to resolve it is adding an optional output
filename for this file, but this would break the uniformity of the
behavior when split-verilog is turn on or off.

Note that this option also works for targets other than
verilog/systemverilog.

This switch does not affect the workflows that exists.

Signed-off-by: unlsycn <unlsycn@unlsycn.com>