[OM] Remove Symbol trait from ClassFieldLikes.
Since 5fb4cf5fd3a3dcaa05b43e340120ce9c3a75f597, ClassLikes are not
SymbolTables, due to the usual reasons that led to us inventing
InnerSymbolTables. However, that change left the Symbol trait on
ClassFieldLikes, which is not legal. We need to fix this, and stop
using Symbols in illegal locations.

This change removes the Symbol trait, and includes some minor juggling
so the current implementation can continue working with minimal
changes.

This renames the attribute name from sym_name to name and adds an
interface method to ClassFieldLike called getNameAttr. Together, these
changes align the accessor for this field to match the previous
getNameAttr accessor on Symbols.

This also updates a couple call sites that directly called
getSymNameAttr to use getNameAttr like the other call sites.

This updates the ObjectFieldOp verifier, which already has to linearly
walk a ClassLike's fields, to stop walking SymbolOpInterface and
instead directly walk ClassFieldLikes.

The last thing is a non-change, where we keep the name attributes as
SymbolNameAttr in ODS, so the printed form still looks like a
symbol. This also aligns with the ObjectFieldOp printed form. This is
so this change can be backwards compatible in the printed IR; if we
think it is better to call a spade a spade, this can be updated to
print these field names as quoted strings, to better reflect that
they're actually just StringAttrs.

Fixes https://github.com/llvm/circt/issues/6659.