[Calyx] Make ControlOp a SymbolTable.
During the Calyx control lowering, in an intermediate state, we have
fsm.machine ops embedded within the calyx.control op. fsm.machine
defines a Symbol, so this is illegal unless calyx.control is a
SymbolTable.

We consistently hit issues when we try to nest SymbolTables, but I
don't think those will actually bite us in the Calyx lowering
pipeline. Since no one is actively maintaining this, I'm proposing to
add SymbolTable to ControlOp until we have a reason to change this, so
the IR is legal during the Calyx lowering.

Fixes https://github.com/llvm/circt/issues/6667.