[FSM] Remove Symbol trait from InstanceOp and HWInstanceOp.
These are used in contexts that are not SymbolTables, so these cannot
be Symbols. It appears that nothing actually needs the Symbol
functionality, and all we need here is the string name. This removes
the Symbol trait and renames the attribute to just `name`.

Fixes https://github.com/llvm/circt/issues/6668.