[Seq] Remove Symbol trait from HLMemOp.
These are used in contexts that are not SymbolTables, so these cannot
be Symbols. It appears that nothing actually needs the Symbol
functionality, and all we need here is the string name. This removes
the Symbol trait and renames the attribute to just name.

If anything needs to reference HLMemOps symbolically, these can become
InnerSymbols, but again, I didn't see anything about that in the
codebase.

Fixes https://github.com/llvm/circt/issues/6661.