[FIRRTL] InferResets: verify that FART annotation is on async resets (#6674)
The FullAsynchronousResetAnnotation is used to mark an asynchronous
reset signal in a module to be used as the reset for all registers that
do not already have an async reset in this module and below in the
hierarchy. There was no error checking that this annotation was attached
to an asynchronous reset typed signal, and the pass would accidentally
try to wire whatever signal was annotated. In some cases, this could
lead to a connecting the signal to an async typed port, which would
become a type checking error in the verifier. This changes InferResets
to verify that the FART annotated signal is in fact asynchronous.