[LowerClasses] Lower classes that instantiate properties. (#6688)
The previous logic would create classes for any FIRRTL modules that
had property ports. However, if some intermediate modules had property
ports, but their parents did not, we would not convert their parents,
which would break pass invariants and lead to failed assertions.

This adds logic to also convert any FIRRTL modules that instantiate
FIRRTL modules that will be converted. This upholds the pass
invariants, and makes it robust in case property ports are used in
part of the hierarchy but not exported from the top-level module of
the circuit.