[HW] HWModule: store input port locations only on block args
On modules we keep an array attribute `port_locs` for port location
information.  On regular HWModules specifically, which have a body
block, we copy the input port location to the entry block arguments, for
error messages and such. This is violating the "single source of truth"
principle for input port location info, so we added a verifier that
ensured the block argument locations were always kept in sync with the
port location attributes. This lead to another problem: if you print our
IR in the generalized format and try to read it back in, by default with
`-mlir-print-debug-info` printing turned off, it will end up throwing
away the location on the blocks. When the IR is read back in, the block
arguments will have new location information corresponding to the
intermediate file, while the location array maintains the old locations.
This trips the verifier since the port attribute location will be out of
sync with the block argument location.

To fix this problem, this changes the HWModule op to record input port
locations only on the blocks, and to store the output port locations in
a new `result_locs` array.  The other module types remain the same.
Now, when printing in the generic format, input ports will lose their
original location information, but the result locations will remain the
same.  If printing with debug information, then both will retain their
original locations.

As an example:
```mlir
// locs.mlir
hw.module @B(in %a: i1, out nameOfPortInSV: i1) {
  hw.output %a: i1
}
```

Printing without the debug information with`circt-opt
-mlir-print-op-generic locs.mlir | ./bin/circt-opt
--mlir-print-debuginfo` gives:
```mlir
module {
  hw.module @B(in %a : i1 loc(#loc2), out nameOfPortInSV : i1 loc(#loc3)) {
    hw.output %a : i1 loc(#loc4)
  } loc(#loc1)
} loc(#loc)
```

Printing with the debug information with `circt-opt
-mlir-print-op-generic --mlir-print-debuginfo locs.mlir |
./bin/circt-opt --mlir-print-debuginfo` gives:
```mlir
module {
  hw.module @B(in %a : i1 loc(#loc2), out nameOfPortInSV : i1 loc(#loc3)) {
    hw.output %a : i1 loc(#loc4)
  } loc(#loc1)
} loc(#loc)
```