[FIRRTL] Add parser version APIs that accept an SMLoc, NFC. (#6692)
These currently default to calling emitError(), which uses the current
token's location. In some cases, the current token may have been
consumed by the time we wish to check versions and potentially emit
errors. This allows users to supply an SMLoc if necessary, and
otherwise defaults to the previous behavior.

None of our current parsing actually consumes a token before checking
versions, but I would like to add a version check to an existing
parser that will come after it has consumed the token.