[ImportVerilog] Add translation, run Slang compilation (#6708)
Add an `--import-verilog` translation target for `circt-translate`. This
facilitates writing single-file test cases that don't require
configuration of the Slang language frontend.

Actually run the Slang elaboration and compilation from within
`ImportContext::importVerilog`. This will essentially do the full
frontend work to ingest Verilog, but the part that translates from the
Slang AST to CIRCT IR ops is still left to be filled in. The import will
always produce an empty MLIR module for now. However, reporting of
syntax or other errors on the Verilog side is already covered and can be
tested for.