[circt-translate] Make ImportVerilog registration conditional
Move the call to `registerFromVerilogTranslation()` from the
`InitAllTranslations.h` header, where it is executed unconditionally,
into the `circt-translate.cpp` file, where it can more easily be guarded
behind an ifdef. This fixes builds with the Slang frontend disabled.

In the future we may want to generated a header with macros that reflect
which optional parts are enabled or disabled. `InitAllTranslations.h`
could then use those macros to conditionally register the ImportVerilog
translation.