[FIRRTL] Support PropertyType in emitConnect.
We have a centralized emitConnect helper that knows how to create the
right kind of connections for various special, non-base types, like
RefType and AnalogType. This adds another special case for
PropertyType, which emits a PropAssignOp.

This is used in various places, but the pass where I noticed this was
missing is InjectDUTHierarchy, so I've added a test there similar to
the test for RefType.