[MSFT][Emit] Replace `output_file` with an `emit::File`
`output_file` is to be replaced with explicit files constructed via the emit dialect.
This PR replaces its use in the MSFT dialect.
Behaviour is slighly altered: the verbatim is always emitted to a file as we should not mix in TCL with SV and it is omitted from the default file list as it should always point to SV.