[HGLDD] Uniquify object names during emission
Ensure the object and variable names emitted by the `EmitHGLDD`
translation are unique and don't collide with any known Verilog
keywords.

Refactor the emission code by pulling most of the shared emission state
into a `GlobalState` struct that is created by the `Emitter` and is
shared by each `FileEmitter`.

Move the `objectNamespace` that is used to uniquify object names into
the new global emitter state, and use `sv::legalizeName` to uniquify
module names instead of `Namespace`. This avoids keyword collisions.
Also add a per-module `moduleNamespace` that is prepopulated with the
global `objectNamespace`, and use it to uniquify variable and instance
names.

Fixes #6749.