[Support] Make isAncestor filterable
Add a function argument to the InstanceGraph::isAncestor member function
to allow this to skip over certain instances.  This is intended to be used
for things like checking if a module is an ancestor of another without
looking through binds.

This is currently not load bearing.

This may help with checks like #5530.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>