[LowerToHW] Do not set output_file if already set
Fix a bug in LowerToHW where a '#hw.output_file' would be incorrectly set
on a module that already has one.  This could result in behavior where an
earlier pass sets up an output structure which is then changed by
LowerToHW due to the location of the module in the instance graph.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>