[Moore] Drop obsolete variable and assign ops
Remove the `moore.mir.vardecl` and `moore.mir.assign` ops which aren't
really used anywhere. We will replace these with improved ops as work on
ImportVerilog progresses.

This also gets rid of the `!moore.lvalue` wrapper type. We'll eventually
want something similar, but we can always create a new version of this
type down the road.