[FIRRTL] Fix Grand Central and Layers Directories
Fix a bug involving interactions of Grand Central and layers output
directories.  When compiling a design that has a DUT and a testbench
directory, treat modules that are instantiated by Grand Central and under
a bound module (which is lowered from a layer) as belonging to Grand
Central and should be placed in its output directory.

This will be revisited in the future as the FIRRTL ABI moves more towards
filelists or manifests.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>