[Calyx] Use a three step rewrite in RemoveCombGroups.
The DialectConversion infrastructure now asserts that replaced ops are
not modified, including inlining the region of a replaced op into a
new op. The patch that added the new assertions explicitly mentioned
instead, we should create a new op, inline the region, and then
replace the old op with the new op.